Changed mongodb url
