+++ BackEnd/.env
@@ -1,2 +1,2 @@
-DB_LINK =  "mongodb+srv://admin:fM7mDS2ATXsQIHtF@clusterzomatoclone.aaj7gl1.mongodb.net/"
+DB_LINK =  "mongodb+srv://admin:uMefkhlKFPy2xUq0@clusterzomatoclone.aaj7gl1.mongodb.net/zomatoclone?retryWrites=true&w=majority&appName=ClusterZomatoClone"
 PORT = 9000
