+++ BackEnd/package.json
@@ -1,13 +1,20 @@
 {
   "dependencies": {
+    "@types/bcrypt": "^5.0.2",
+    "bcrypt": "^5.1.1",
+    "bcrypt-ts": "^5.0.2",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "email-validator": "^2.0.4",
     "express": "^4.19.1",
+    "fix-esm": "^1.0.1",
     "lodash": "^4.17.21",
-    "mongoose": "^8.2.3"
+    "mongodb": "^6.5.0",
+    "mongoose": "^8.2.3",
+    "tsconfig": "^7.0.0"
   },
   "devDependencies": {
+    "@tsconfig/recommended": "^1.0.5",
     "@types/express": "^4.17.21",
     "@types/node": "^20.11.30",
     "nodemon": "^3.1.0",
