+++ BackEnd/src/models/DeliveryPerson/index.ts
@@ -1,0 +1,1 @@
+export {default} from './DeliveryPerson'
