+++ BackEnd/src/models/RestrauntOwner/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./RestrauntOwner";
