+++ BackEnd/src/models/User/index.ts
@@ -1,0 +1,2 @@
+const userModel = require('./User')
+export = userModel;
