+++ BackEnd/tsconfig.json
@@ -1,11 +1,16 @@
 {
+  "extends": "@tsconfig/recommended/tsconfig.json",
   "compilerOptions": {
     "lib": ["es6"],
     "target": "es6",
-    "module": "commonjs",
-    "rootDir": "./",
-    "outDir": "./build",
+    "module": "CommonJS",
     "esModuleInterop": true,
-    "strict": true
-  }
+    "strict": true,
+    "outDir": "./build",
+    "baseUrl": "./src",
+    "paths": {
+      "*": ["*"]
+    }
+  },
+  "include": ["src"]
 }
